+++ alf/utils/common.py
@@ -65,34 +65,6 @@ def orig_tf_gfile_context():
         tf.io.gfile = TB_IO_GFILE
 
 
-@contextmanager
-def original_torch_module_functions():
-    """A context manager for restoring some key original nn.Module functions that
-    have been overwritten by ALF.
-
-    This can be used when we are trying to load a pretrained huggingface model which
-    require a newer and original ``torch.nn.Module``.
-
-    Example:
-
-    .. code-block:: python
-
-        with original_torch_module_functions():
-            model = hf_model_load()
-    """
-    keys = [
-        'state_dict', 'load_state_dict', '_save_to_state_dict',
-        '_load_from_state_dict'
-    ]
-    current_funcs = {k: getattr(nn.Module, k) for k in keys}
-    old_funcs = {k: getattr(alf_module, 'old_' + k) for k in keys}
-    for k in keys:
-        setattr(nn.Module, k, old_funcs[k])
-    yield
-    for k in keys:
-        setattr(nn.Module, k, current_funcs[k])
-
-
 def add_method(cls):
     """A decorator for adding a method to a class (cls).
     Example usage:
